+++ index.ts
@@ -11,5 +11,6 @@ export { PageScrollIndicator } from "./src/PageScrollIndicator";
 export { Rating } from "./src/Rating";
 export { SkipLink } from "./src/SkipLink";
 export { Slider } from "./src/Slider";
+export { TextReveal } from "./src/TextReveal";
 export { Tabs } from "./src/Tabs";
 export { Timeline } from "./src/Timeline";