+++ src/SkipLink/SkipLink.tsx
@@ -5,7 +5,7 @@ export const SkipLink = ({
   text = "Skip to content",
   targetId = "main",
   className,
-  ...props
+  ...rest
 }: SkipLinkProps) => {
   return (
     <a
@@ -14,7 +14,7 @@ export const SkipLink = ({
         "w-fit absolute visually-hidden top-4 left-4 z-50 bg-slate-200 dark:bg-slate-900 px-4 py-2 text-slate-950 dark:text-slate-200 font-medium outline-none focus-visible:outline-sky-700 dark:focus-visible:outline-amber-400 focus-visible:outline-2 focus-visible:outline-offset-4 dark:focus-visible:outline-offset-4",
         className
       )}
-      {...props}
+      {...rest}
     >
       {text}
     </a>